requirements.txt
pandas
numpy
openai
python-dotenv
fastapi[all]